% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{FizzBuzz}
\alias{FizzBuzz}
\title{FizzBuzz}
\usage{
FizzBuzz(n)
}
\arguments{
\item{n}{A positive integer}
}
\value{
A vector of numbers will be printed. In the vector, if the number is divisible by 3 "Fizz" will appear,
if the number is divisible by 5 "Buzz" will appear,
and if the number is divisible by both 3 and 5 "FizzBuzz" will appear.
}
\description{
For any number evenly divisible by 3, there will be "Fizz" instead of a number.
The same will be done for any number evenly divisible by 5 except "Buzz" will replace the number.
If the integer is divisible by 3 and 5, the number will be replaced with "FizzBuzz".
}
\examples{
FizzBuzz(10)
FizzBuzz(30)

}
