% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{Flagstaff_Weather}
\alias{Flagstaff_Weather}
\title{A time series of weather in Flagstaff, AZ.}
\format{
A data frame with 25,449 observations with 5 columns.
\describe{
   \item{Date}{The dates of each individual observation.}
   \item{TMIN}{Temperature minimum of each date observed.}
   \item{TMAX}{Temperature maximum of each date observed.}
   \item{Precipitation}{The amount of percipitation on each date observed.}
   \item{SNOW}{The amount of snow on each date observed.}
}
}
\usage{
Flagstaff_Weather
}
\description{
A set of data about Flagstaff's weather from 1950 to 2019.
}
\references{
https://www.ncdc.noaa.gov/cdo-web/search
}
\keyword{datasets}
